+++ firmware/application/CMakeLists.txt
@@ -282,6 +282,8 @@ set(CPPSRC
 	apps/tpms_app.cpp
 	apps/tpms_app.cpp
 	apps/ui_spectrum_painter.cpp
+	apps/ui_spectrum_painter_image.cpp
+	apps/ui_spectrum_painter_text.cpp
 	protocols/aprs.cpp
 	protocols/ax25.cpp
 	protocols/bht.cpp
